requirements.txt
torch
torchvision
transformers
Pillow
huggingface_hub
safetensors
